# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZenKart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      summary: Apply coupon to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyCouponBody"
      responses:
        "200":
          description: Coupon applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
...
  /chat/conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /chat/start:
    post:
      operationId: startConversation
      tags: [chat]
      summary: Start a conversation about a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationBody"
      responses:
        "201":
          description: Conversation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  # DASHBOARD SUMMARY
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [buyer, seller, reseller, used_goods_seller, admin]
        avatarUrl:
          type: string
          nullable: true
        walletBalance:
          type: number
        zenCoins:
          type: integer
        isVerified:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, walletBalance, zenCoins, isVerified, createdAt]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        mrp:
          type: number
        discountPercent:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        brand:
          type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isFlashSale:
          type: boolean
        badge:
          type: string
          nullable: true
        deliveryDays:
          type: integer
        createdAt:
          type: string
      required: [id, title, price, mrp, discountPercent, images, categoryId, sellerId, brand, stock, rating, reviewCount, isFeatured, isFlashSale, createdAt]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        selectedVariant:
          type: string
          nullable: true
        savedForLater:
          type: boolean
      required: [id, productId, product, quantity, savedForLater]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        couponDiscount:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        appliedCoupon:
          type: string
          nullable: true
        itemCount:
          type: integer
      required: [items, subtotal, discount, couponDiscount, deliveryCharge, total, itemCount]
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        selectedVariant:
          type: string
      required: [productId, quantity]
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
        savedForLater:
          type: boolean
      required: []
    ApplyCouponBody:
      type: object
      properties:
        couponCode:
          type: string
      required: [couponCode]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        text:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, conversationId, senderId, senderName, text, isRead, createdAt]
    Conversation:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
          nullable: true
        listingTitle:
          type: string
          nullable: true
        listingImage:
          type: string
          nullable: true
        otherUserId:
          type: integer
        otherUserName:
          type: string
        otherUserAvatar:
          type: string
          nullable: true
        lastMessage:
          type: string
          nullable: true
        unreadCount:
          type: integer
        updatedAt:
          type: string
      required: [id, otherUserId, otherUserName, unreadCount, updatedAt]
    SendMessageBody:
      type: object
      properties:
        text:
          type: string
      required: [text]
    StartConversationBody:
      type: object
      properties:
        listingId:
          type: integer
        sellerId:
          type: integer
        initialMessage:
          type: string
      required: [listingId, sellerId, initialMessage]
...